+++ pylab/VERSION
@@ -1,1 +1,1 @@
-0.8.2
+0.9.0
