+++ pylab/io/__init__.py
@@ -1,0 +1,1 @@
+from .writer import CustomWriter
