+++ pylab/maingui.py
@@ -1,11 +1,4 @@
-import datetime
-from itertools import count
-from pathlib import Path
-import numpy as np
-import pandas as pd
-
 from pymmcore_plus import CMMCorePlus
-from useq import MDAEvent
 
 # Necessary modules for the IPython console
 from qtconsole.rich_jupyter_widget import RichJupyterWidget
@@ -51,28 +44,13 @@ class MainWindow(QMainWindow):
         toggle_console_action = self.menuBar().addAction("Toggle Console")
         toggle_console_action.triggered.connect(self.toggle_console)
 
+        self.thor_gui.mmc.mda.events.sequenceFinished.connect(self._on_end)
         self.cfg_gui.configUpdated.connect(self._update_config)
         self.cfg_gui.recordStarted.connect(self.record)
         
     def record(self):
-        self.thor_gui.mda.run_mda()   
         self.dhyana_gui.mda.run_mda()
-         
-    def _update_config(self, config):
-        self.config: ExperimentConfig = config
-        self._refresh_mda_gui()
-        self._refresh_save_gui()
-        
-    def _refresh_mda_gui(self):
-        self.dhyana_gui.mda.setValue(self.config.meso_sequence)
-        self.thor_gui.mda.setValue(self.config.pupil_sequence)
-        
-    def _refresh_save_gui(self):
-        self.dhyana_gui.mda.save_info.setValue({'save_dir': str(self.config.bids_dir),  'save_name': str(self.config.meso_file_path), 'format': 'ome-tiff', 'should_save': True})
-        self.thor_gui.mda.save_info.setValue({'save_dir': str(self.config.bids_dir), 'save_name': str(self.config.pupil_file_path), 'format': 'ome-tiff', 'should_save': True})
-        
-    def _on_pause(self, state: bool) -> None:
-        """Called when the MDA is paused."""
+        self.thor_gui.mda.run_mda()   
         
     def toggle_console(self):
         """Show or hide the IPython console."""
@@ -125,4 +103,23 @@ class MainWindow(QMainWindow):
             # Optional, so you can use 'self' directly in the console
         })
 
+    def _on_end(self) -> None:
+        """Called when the MDA is finished."""
+        self.cfg_gui.save_config()
+        self.plots()
 
+    def _update_config(self, config):
+        self.config: ExperimentConfig = config
+        self._refresh_mda_gui()
+        self._refresh_save_gui()
+        
+    def _refresh_mda_gui(self):
+        self.dhyana_gui.mda.setValue(self.config.meso_sequence)
+        self.thor_gui.mda.setValue(self.config.pupil_sequence)
+        
+    def _refresh_save_gui(self):
+        self.dhyana_gui.mda.save_info.setValue({'save_dir': str(self.config.bids_dir),  'save_name': str(self.config.meso_file_path), 'format': 'ome-tiff', 'should_save': True})
+        self.thor_gui.mda.save_info.setValue({'save_dir': str(self.config.bids_dir), 'save_name': str(self.config.pupil_file_path), 'format': 'ome-tiff', 'should_save': True})
+        
+    def _on_pause(self, state: bool) -> None:
+        """Called when the MDA is paused."""
