+++ pylab/mmcore.py
@@ -1,25 +1,24 @@
 import pymmcore_plus
 
-DHYANA_CONFIG = r'C:/Program Files/Micro-Manager-2.0/mm-sipefield.cfg'
-THOR_CONFIG = r'C:/Program Files/Micro-Manager-2.0/ThorCam.cfg'
-
 import logging
 
-import pylab
 import useq
-from pylab.engine import MesoEngine, PupilEngine
+import logging
 
 # Disable pymmcore-plus logger
 package_logger = logging.getLogger('pymmcore-plus')
+
 # Set the logging level to CRITICAL to suppress lower-level logs
 package_logger.setLevel(logging.CRITICAL)
+
+DHYANA_CONFIG = r'C:/Program Files/Micro-Manager-2.0/mm-sipefield.cfg'
+THOR_CONFIG = r'C:/Program Files/Micro-Manager-2.0/ThorCam.cfg'
 
 def load_cores():
     mmcore_dhyana: pymmcore_plus.CMMCorePlus = pymmcore_plus.CMMCorePlus(mm_path=r'C:\Program Files\Micro-Manager-2.0')
     mmcore_thor: pymmcore_plus.CMMCorePlus = pymmcore_plus.CMMCorePlus(mm_path=r'C:\Program Files\Micro-Manager-thor')
 
     mmcore_dhyana.loadSystemConfiguration(DHYANA_CONFIG)
-
     mmcore_thor.loadSystemConfiguration(THOR_CONFIG)
 
     mmcore_dhyana.setCircularBufferMemoryFootprint(10000)
@@ -29,13 +28,18 @@ def load_cores():
     return mmcore_dhyana, mmcore_thor
 
 def test_mda(frames: int = 20000):    
-    from pymmcore_plus.mda.handlers import OMETiffWriter, ImageSequenceWriter
-    from pylab.writer import CustomWriter
+    #from pymmcore_plus.mda.handlers import OMETiffWriter, ImageSequenceWriter
+    from pylab.engines import MesoEngine, PupilEngine
+    from pylab.io import CustomWriter
+    
     mmcore_dhyana, mmcore_thor = load_cores()
+    
     load_dhyana_mmc_params(mmcore_dhyana)
     load_thorcam_mmc_params(mmcore_thor)
+    
     mmcore_dhyana.mda.set_engine(MesoEngine(mmcore_dhyana, use_hardware_sequencing=True)) 
     mmcore_thor.mda.set_engine(PupilEngine(mmcore_thor, use_hardware_sequencing=True))
+    
     print('running MDA test...')
     mmcore_dhyana.run_mda(useq.MDASequence(time_plan={"interval": 0, "loops": frames}), output=CustomWriter(r'C:/dev/dh.ome.tif'))
     mmcore_thor.run_mda(useq.MDASequence(time_plan={"interval": 0, "loops": frames}), output=CustomWriter(r'C:/dev/thor.ome.tif'))
@@ -83,5 +87,5 @@ def load_dev_cores() -> pymmcore_plus.CMMCorePlus:
     core2.loadSystemConfiguration()
     return core1, core2
     
-if __name__ == '__main__':
-    test_mda(100)
+# if __name__ == '__main__':
+#     test_mda(100)
