+++ pylab/widgets/__init__.py
@@ -1,0 +1,3 @@
+from .controller import ConfigController
+from .mdagui import CustomMDAWidget, MDA
+from .viewer import ImagePreview
