+++ pylab/widgets/mdagui.py
@@ -1,0 +1,86 @@
+from pymmcore_plus import CMMCorePlus
+
+from PyQt6.QtWidgets import (
+    QGroupBox,
+    QHBoxLayout,
+    QVBoxLayout,
+    QWidget,
+)
+
+from pymmcore_widgets import (
+    MDAWidget,
+    ExposureWidget,
+    LiveButton,
+    SnapButton,
+)
+
+from pylab.config import ExperimentConfig
+from pylab.io.writer import CustomWriter
+from .viewer import ImagePreview
+
+class CustomMDAWidget(MDAWidget):
+    def run_mda(self) -> None:
+        """Run the MDA sequence experiment."""
+        # in case the user does not press enter after editing the save name.
+        self.save_info.save_name.editingFinished.emit()
+
+
+        sequence = self.value()
+
+        # technically, this is in the metadata as well, but isChecked is more direct
+        if self.save_info.isChecked():
+            save_path = self._update_save_path_from_metadata(
+                sequence, update_metadata=True
+            )
+        else:
+            save_path = None
+
+        # run the MDA experiment asynchronously
+        self._mmc.run_mda(sequence, output=CustomWriter(save_path))
+
+class MDA(QWidget):
+    """An example of using the MDAWidget to create and acquire a useq.MDASequence.
+
+    The `MDAWidget` provides a GUI to construct a `useq.MDASequence` object.
+    This object describes a full multi-dimensional acquisition;
+    In this example, we set the `MDAWidget` parameter `include_run_button` to `True`,
+    meaning that a `run` button is added to the GUI. When pressed, a `useq.MDASequence`
+    is first built depending on the GUI values and is then passed to the
+    `CMMCorePlus.run_mda` to actually execute the acquisition.
+    For details of the corresponding schema and methods, see
+    https://github.com/pymmcore-plus/useq-schema and
+    https://github.com/pymmcore-plus/pymmcore-plus.
+
+    """
+
+    def __init__(self, core_object: CMMCorePlus, cfg) -> None:
+        super().__init__()
+        # get the CMMCore instance and load the default config
+        self.mmc = core_object
+        self.config: ExperimentConfig = cfg
+
+        # instantiate the MDAWidget
+        self.mda = CustomMDAWidget(mmcore=self.mmc)
+        # ----------------------------------Auto-set MDASequence and save_info----------------------------------#
+        self.mda.setValue(self.config.pupil_sequence)
+        self.mda.save_info.setValue({'save_dir': r'C:/dev', 'save_name': 'file', 'format': 'ome-tiff', 'should_save': True})
+        # -------------------------------------------------------------------------------------------------------#
+        self.setLayout(QHBoxLayout())
+
+        self.preview = ImagePreview(mmcore=self.mmc,parent=self.mda)
+        self.snap_button = SnapButton(mmcore=self.mmc)
+        self.live_button = LiveButton(mmcore=self.mmc)
+        self.exposure = ExposureWidget(mmcore=self.mmc)
+
+        live_viewer = QGroupBox()
+        live_viewer.setLayout(QVBoxLayout())
+        buttons = QGroupBox()
+        buttons.setLayout(QHBoxLayout())
+        buttons.layout().addWidget(self.snap_button)
+        buttons.layout().addWidget(self.live_button)
+        live_viewer.layout().addWidget(buttons)
+        live_viewer.layout().addWidget(self.preview)
+
+        self.layout().addWidget(self.mda)
+        self.layout().addWidget(live_viewer)
+
